move context/overflow menu to the main MDP header bar
