+++ packages/veritone-react-common/src/components/MediaInfoPanel/styles.scss
@@ -13,30 +13,6 @@
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
-
-    .headerMenu {
-      display: flex;
-      flex-flow: row;
-
-      .popperClose {
-        pointer-events: none;
-      }
-
-      .poperOpen {
-        margin-right: 20px;
-      }
-
-      .closeButton {
-        font-size: 24px;
-      }
-
-      .headerMenuItem {
-        color: $dark-black;
-        font-family: Roboto;
-        font-size: 14px;
-        line-height: 16px;
-      }
-    }
   }
 
   .infoPanelContent {
